pyproject.toml
[project]
name = "mimiod"
version = "0.1.0"
description = "MongoDB database generation from natural language using agentic LLM workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic[email]>=2.0.0",
    "pymongo>=4.0.0",
    "faker>=20.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
